+++ close.go
@@ -1,6 +1,9 @@
-package ext
+package serverz
 
-import "github.com/goph/serverz/internal/errors"
+import (
+	"errors"
+	"fmt"
+)
 
 // NoopCloser is a dummy Closer implementation which can be used as a fallback.
 var NoopCloser = CloserFunc(func() {})
@@ -16,12 +19,23 @@ type CloserFunc func()
 // Close calls the underlying function and converts any panic to an error.
 func (f CloserFunc) Close() (err error) {
 	defer func() {
-		err = errors.Recover(recover())
+		r := recover()
+
+		if r != nil {
+			switch x := r.(type) {
+			case string:
+				err = errors.New(x)
+			case error:
+				err = x
+			default:
+				err = fmt.Errorf("Unknown panic, received: %v", r)
+			}
+		}
 	}()
 
 	f()
 
-	return err
+	return
 }
 
 // Closers is a collection of Closer instances.
@@ -33,13 +47,24 @@ func (c Closers) Close() error {
 		return nil
 	}
 
-	errBuilder := errors.NewMultiErrorBuilder()
+	var merr multiError
 
 	for _, closer := range c {
 		err := closer.Close()
 
-		errBuilder.Add(err)
+		if err != nil {
+			merr = append(merr, err)
+		}
 	}
 
-	return errBuilder.ErrOrNil()
+	return merr.ErrOrNil()
+}
+
+// Close calls Close method on a struct if it implements the Closer interface.
+func Close(closer interface{}) error {
+	if c, ok := closer.(Closer); ok {
+		return c.Close()
+	}
+
+	return nil
 }