+++ internal/named.go
@@ -1,7 +1,0 @@
-package internal
-
-// NamedServer exposes a name for a server.
-type NamedServer interface {
-	// GetName returns the name of a server.
-	GetName() string
-}