+++ server.go
@@ -26,3 +26,10 @@ type Server interface {
 	// For a graceful shutdown, use Shutdown.
 	Close() error
 }
+
+// ListenServer can listen on an interface on its own when provided with an address.
+type ListenServer interface {
+	// ListenAndServe listens on a network address and then
+	// calls Serve to handle requests on incoming connections.
+	ListenAndServe(addr net.Addr) error
+}